chore: Switched to auth atoms
