+++ components/Navbar/HelpLink.tsx
@@ -10,16 +10,17 @@ import Link from 'next/link'
 import { Tooltip } from 'ui'
 import { themeSelector } from 'lib/themeSelector'
 import { getCurrentLessonKey, getChapterKey, keys } from 'lib/progress'
-import { useAuthContext } from 'contexts/AuthContext'
 import { usePathname } from 'next/navigation'
 import clsx from 'clsx'
+import { useAtom } from 'jotai'
+import { accountAtom } from 'state/state'
 
 export default function HelpLink({ params }: { params: any }) {
   const lang = useLang()
   const t = useTranslations(lang)
   const { chaptersUrl } = useLocalizedRoutes()
   const { chapterId } = usePathData()
-  const { account } = useAuthContext()
+  const [account] = useAtom(accountAtom)
   const pathName = usePathname() || ''
 
   const pathData = pathName.split('/').filter((p) => p)
