+++ content/lessons/chapter-4/public-key-3.tsx
@@ -6,9 +6,10 @@ import { useTranslations } from 'hooks'
 import { Text } from 'ui'
 import { useState } from 'react'
 import { getLessonKey } from 'lib/progress'
-import { useAuthContext } from 'contexts/AuthContext'
 import { useDataContext } from 'contexts/DataContext'
 import { getLanguageString } from 'lib/SavedCode'
+import { useAtom } from 'jotai'
+import { accountAtom } from 'state/state'
 
 export const metadata = {
   title: 'chapter_four.public_key_three.title',
@@ -17,7 +18,7 @@ export const metadata = {
 
 export default function PublicKey3({ lang }) {
   const t = useTranslations(lang)
-  const { account } = useAuthContext()
+  const [account] = useAtom(accountAtom)
   const { currentLanguage } = useDataContext()
   const [privateKey, setPrivateKey] = useState('')
