+++ content/lessons/chapter-7/mempool-transaction-1.tsx
@@ -6,9 +6,10 @@ import { useTranslations } from 'hooks'
 import { Text } from 'ui'
 import { useState } from 'react'
 import { getLessonKey } from 'lib/progress'
-import { useAuthContext } from 'contexts/AuthContext'
 import { useDataContext } from 'contexts/DataContext'
 import { getLanguageString } from 'lib/SavedCode'
+import { useAtom } from 'jotai'
+import { accountAtom } from 'state/state'
 
 export const metadata = {
   title: 'chapter_seven.mempool_transaction_one.title',
@@ -91,7 +92,7 @@ export default function MempoolTransaction1({ lang }) {
     ],
   ]
 
-  const { account } = useAuthContext()
+  const [account] = useAtom(accountAtom)
 
   const javascript = {
     program: `//BEGIN VALIDATION BLOCK
