+++ state/state.ts
@@ -1,0 +1,5 @@
+import { atom } from 'jotai'
+
+export const accountAtom = atom<any | undefined>(undefined)
+
+export const isAuthLoadingAtom = atom<boolean>(true)
